+++ backend/views/review.py
@@ -1,21 +1,23 @@
 from flask import Flask, request, jsonify, Blueprint
+from flask_jwt_extended import get_jwt_identity, jwt_required
 from models import db, Review
 
 review_bp = Blueprint("review_blueprint", __name__)
 
-@review_bp.route("/restaurants/<restaurant_id>/reviews", methods=["POST"])
+@review_bp.route("/restaurants/<int:restaurant_id>/reviews", methods=["POST"])
+@jwt_required()
 def add_review(restaurant_id):
+    current_user = get_jwt_identity()
     data = request.get_json()
 
-    user_id = data.get('user_id')
     rating = data.get('rating')
     comment = data.get('comment')
 
-    if not all([user_id, rating, comment]):
+    if not all([rating, comment]):
         return jsonify({"error": "User ID, Rating, and Comment are required!"}), 400
 
     new_review = Review(
-        user_id=user_id,
+        user_id=current_user,
         restaurant_id=restaurant_id,
         rating=rating,
         comment=comment
@@ -26,7 +28,7 @@ def add_review(restaurant_id):
 
     return jsonify({"message": "Review added successfully!"}), 201
 
-@review_bp.route("/restaurants/<restaurant_id>/reviews", methods=["GET"])
+@review_bp.route("/restaurants/<int:restaurant_id>/reviews", methods=["GET"])
 def get_reviews(restaurant_id):
     reviews = Review.query.filter_by(restaurant_id=restaurant_id).all()
 
@@ -42,12 +44,17 @@ def get_reviews(restaurant_id):
 
     return jsonify(review_list), 200
 
-@review_bp.route("/reviews/<review_id>", methods=["PATCH"])
+@review_bp.route("/reviews/<int:review_id>", methods=["PATCH"])
+@jwt_required()
 def update_review(review_id):
+    current_user = get_jwt_identity()
     review = Review.query.get(review_id)
 
     if not review:
         return jsonify({"error": "Review not found"}), 404
+    
+    if review.user_id != current_user:
+        return jsonify({"error": "Unauthorised access, log in first!"})  
 
     data = request.get_json()
 
@@ -58,12 +65,17 @@ def update_review(review_id):
 
     return jsonify({"message": "Review updated successfully!"}), 200
 
-@review_bp.route("/reviews/<review_id>", methods=["DELETE"])
+@review_bp.route("/reviews/<int:review_id>", methods=["DELETE"])
+@jwt_required()
 def delete_review(review_id):
+    current_user = get_jwt_identity()
     review = Review.query.get(review_id)
 
     if not review:
         return jsonify({"error": "Review not found"}), 404
+    
+    if review.user_id != current_user:
+        return jsonify({"error": "Unauthorised access, log in first!"})
 
     db.session.delete(review)
     db.session.commit()
